+++ tests/wycheproof/test_ecdsa.py
@@ -67,10 +67,7 @@ def test_ecdsa_signature(backend, wycheproof):
             ),
         )
         assert isinstance(key, ec.EllipticCurvePublicKey)
-    except (UnsupportedAlgorithm, ValueError):
-        # In some OpenSSL 1.1.1 versions (RHEL and Fedora), some keys fail to
-        # load with ValueError, instead of  Unsupported Algorithm. We can
-        # remove handling for that exception when we drop support.
+    except UnsupportedAlgorithm:
         pytest.skip(
             "unable to load key (curve {})".format(
                 wycheproof.testgroup["key"]["curve"]